pyproject.toml
[project]
name = "mafia-bot-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.20.0",
    "bcrypt>=4.2.1",
    "cryptography>=44.0.0",
    "fastapi>=0.115.6",
    "passlib>=1.7.4",
    "python-jose>=3.3.0",
    "python-multipart>=0.0.20",
    "sqlalchemy>=2.0.36",
    "uvicorn>=0.34.0",
]
